first set of updates
made some updates. realized the mobile menu is visible if you scroll down, so that needs to be hidden better. Tried display none and z-index -1 but neither works. display none might work if I can add a delay between the transform and the display none so it can still slide in and out of the screen.